fix: more display in navtab
